requirements.txt
# Databricks connectivity
databricks-sql-connector>=3.0.0
databricks-sdk>=0.18.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LLM and NLP
mistralai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Security and validation
sqlparse>=0.4.4
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Rate limiting
ratelimit>=2.2.1

# Web UI
streamlit>=1.28.0

# Delta Lake support
delta-spark>=2.4.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
